# openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "EduSearch API",
    "description": "Comprehensive Indian Education Data API — JEE/NEET cutoffs, NIRF rankings, college data, placements, and admission predictions.",
    "version": "1.0.0",
    "contact": {
      "name": "EduSearch API"
    }
  },
...
  "paths": {
...
    "/api/v1/rankings/nirf": {
      "get": {
        "summary": "List NIRF rankings",
        "tags": [
          "Rankings"
        ],
        "description": "Returns NIRF rankings with the overall score and every published sub-score (teaching, research, graduation outcomes, outreach, perception), joined with institute details. Omit `year` to get the most recent ranking year on record. Results are ordered by rank ascending.",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "year",
            "required": false,
            "description": "Ranking year. Defaults to the latest year available."
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "engineering",
                "medical",
                "overall",
                "pharmacy",
                "management"
              ]
            },
            "in": "query",
            "name": "category",
            "required": false,
            "description": "Ranking category to return."
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50
            },
            "in": "query",
            "name": "limit",
            "required": false,
            "description": "Maximum records to return (1-200)."
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            },
            "in": "query",
            "name": "offset",
            "required": false,
            "description": "Records to skip, for paging through results."
          }
        ],
        "responses": {
          "200": {
            "description": "Matching NIRF rankings with pagination metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Matching NIRF rankings with pagination metadata.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always true on a successful response."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "Unique ranking record ID."
                          },
                          "institute_id": {
                            "type": "integer",
                            "description": "ID of the ranked institute."
                          },
                          "year": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Ranking year."
                          },
                          "category": {
                            "type": "string",
                            "nullable": true,
                            "description": "Ranking category: engineering, medical, overall, pharmacy or management."
                          },
                          "rank": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Position in the ranking for that year and category."
                          },
                          "score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Overall NIRF score out of 100."
                          },
                          "tlr_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Teaching, Learning and Resources sub-score."
                          },
                          "rpc_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Research and Professional Practice sub-score."
                          },
                          "go_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Graduation Outcomes sub-score."
                          },
                          "oi_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Outreach and Inclusivity sub-score."
                          },
                          "perception_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Perception sub-score."
                          },
                          "institute_name": {
                            "type": "string",
                            "nullable": true,
                            "description": "Full institute name, joined in for convenience."
                          },
                          "short_name": {
                            "type": "string",
                            "nullable": true,
                            "description": "Institute abbreviation, joined in for convenience."
                          },
                          "type": {
                            "type": "string",
                            "nullable": true,
                            "description": "Institute category, joined in for convenience."
                          },
                          "state": {
                            "type": "string",
                            "nullable": true,
                            "description": "Institute state, joined in for convenience."
                          }
                        }
                      },
                      "description": "The matching records."
                    },
                    "meta": {
                      "type": "object",
                      "description": "Pagination metadata.",
                      "properties": {
                        "total": {
                          "type": "integer",
                          "description": "Total records matching the filters, ignoring limit and offset."
                        },
                        "limit": {
                          "type": "integer",
                          "description": "Maximum records returned in this response."
                        },
                        "offset": {
                          "type": "integer",
                          "description": "Number of records skipped."
                        },
                        "has_more": {
                          "type": "boolean",
                          "description": "True when more records remain beyond this page."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "One or more query parameters were invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "One or more query parameters were invalid.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always false on an error response."
                    },
                    "error": {
                      "type": "string",
                      "description": "Human-readable description of what went wrong."
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code, repeated in the body for convenience."
                    },
                    "details": {
                      "type": "array",
                      "nullable": true,
                      "description": "Per-field validation problems, present on 400 responses.",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Request did not reach the API through the RapidAPI proxy. Only returned by the hosted deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Request did not reach the API through the RapidAPI proxy. Only returned by the hosted deployment.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always false on an error response."
                    },
                    "error": {
                      "type": "string",
                      "description": "Human-readable description of what went wrong."
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code, repeated in the body for convenience."
                    },
                    "details": {
                      "type": "array",
                      "nullable": true,
                      "description": "Per-field validation problems, present on 400 responses.",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/exams/{exam}/stats": {
      "get": {
        "summary": "Get exam statistics by year",
        "tags": [
          "Exams"
        ],
        "description": "Returns registration, attendance and qualification counts plus score ranges for one exam, newest year first. Useful for showing how competitive a given year was. Add `?year=` to pin a single year.",
        "parameters": [
          {
            "schema": {
              "type": "integer"
            },
            "in": "query",
            "name": "year",
            "required": false,
            "description": "Restrict to a single exam year, e.g. 2025."
          },
          {
            "schema": {
              "type": "string",
              "enum": [
                "jee_main",
                "jee_advanced",
                "neet"
              ]
            },
            "in": "path",
            "name": "exam",
            "required": true,
            "description": "Exam to fetch statistics for."
          }
        ],
        "responses": {
          "200": {
            "description": "Exam statistics for the requested exam.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Exam statistics for the requested exam.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always true on a successful response."
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "additionalProperties": true,
                        "properties": {
                          "id": {
                            "type": "integer",
                            "description": "Unique exam statistic record ID."
                          },
                          "exam": {
                            "type": "string",
                            "nullable": true,
                            "description": "Exam identifier: jee_main, jee_advanced or neet."
                          },
                          "year": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Exam year."
                          },
                          "total_registered": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Candidates who registered."
                          },
                          "total_appeared": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Candidates who sat the exam."
                          },
                          "total_qualified": {
                            "type": "integer",
                            "nullable": true,
                            "description": "Candidates who qualified."
                          },
                          "max_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Highest score recorded."
                          },
                          "min_qualifying_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Minimum qualifying score."
                          },
                          "avg_score": {
                            "type": "number",
                            "nullable": true,
                            "description": "Average score across candidates."
                          }
                        }
                      },
                      "description": "Statistics per year, newest first."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "The exam name or year was invalid.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "The exam name or year was invalid.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always false on an error response."
                    },
                    "error": {
                      "type": "string",
                      "description": "Human-readable description of what went wrong."
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code, repeated in the body for convenience."
                    },
                    "details": {
                      "type": "array",
                      "nullable": true,
                      "description": "Per-field validation problems, present on 400 responses.",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "Request did not reach the API through the RapidAPI proxy. Only returned by the hosted deployment.",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Request did not reach the API through the RapidAPI proxy. Only returned by the hosted deployment.",
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "description": "Always false on an error response."
                    },
                    "error": {
                      "type": "string",
                      "description": "Human-readable description of what went wrong."
                    },
                    "statusCode": {
                      "type": "integer",
                      "description": "HTTP status code, repeated in the body for convenience."
                    },
                    "details": {
                      "type": "array",
                      "nullable": true,
                      "description": "Per-field validation problems, present on 400 responses.",
                      "items": {
                        "type": "object",
                        "additionalProperties": true
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    }
  ],
...
}